requirements.txt
fastmcp
httpx
psutil
keyring
pyperclip
numpy
matplotlib
dnspython
pillow